requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
asyncpg>=0.27.0 
pytz>=2023.3
aiohttp>=3.8.0
APScheduler==3.6.3
python-dateutil>=2.8.2
requests>=2.31.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
gspread>=5.12.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.118.0 